Module google_realtimebidding1::api

Structs

ActivatePretargetingConfigRequest
A request to activate a pretargeting configuration. Sets the configuration's state to ACTIVE.

AdTechnologyProviders
The list of detected Ad Technology Providers for this creative. Bids placed for inventory that will serve to EEA or UK users are expected to comply with GDPR requirements. You must ensure that the creatives used in such bids should contain only user consented ad technology providers as indicated in the bid request. Google reserves the right to filter non-compliant bids. User consented ad technology providers can be found in the Google Protocol with the BidRequest.adslot.consented_providers_settings field, and can be found as an OpenRTB extension with the BidRequest.user.ext.consented_providers_settings and BidRequest.user.ext.consent fields. See https://support.google.com/authorizedbuyers/answer/9789378 for additional information about the Google TCF v2 integration.

AddTargetedAppsRequest
A request to start targeting the provided app IDs in a specific pretargeting configuration. The pretargeting configuration itself specifies how these apps are targeted. in PretargetingConfig.appTargeting.mobileAppTargeting.

AddTargetedPublishersRequest
A request to start targeting the provided publishers in a specific pretargeting configuration. The pretargeting configuration itself specifies how these publishers are targeted in PretargetingConfig.publisherTargeting.

AddTargetedSitesRequest
A request to start targeting the provided sites in a specific pretargeting configuration. The pretargeting configuration itself specifies how these sites are targeted in PretargetingConfig.webTargeting.

AdvertiserAndBrand
Detected advertiser and brand information.

AppTargeting
A subset of app inventory to target. Bid requests that match criteria in at least one of the specified dimensions will be sent.

Bidder
Bidder settings.

BidderCreativeListCall
Lists creatives.

BidderCreativeWatchCall
Watches all creatives pertaining to a bidder. It is sufficient to invoke this endpoint once per bidder. A Pub/Sub topic will be created and notifications will be pushed to the topic when any of the bidder's creatives change status. All of the bidder's service accounts will have access to read from the topic. Subsequent invocations of this method will return the existing Pub/Sub configuration.

BidderEndpointGetCall
Gets a bidder endpoint by its name.

BidderEndpointListCall
Lists all the bidder's endpoints.

BidderGetCall
Gets a bidder account by its name.

BidderListCall
Lists all the bidder accounts that belong to the caller.

BidderMethods
A builder providing access to all methods supported on bidder resources. It is not used directly, but through the RealTimeBidding hub.

BidderPretargetingConfigActivateCall
Activates a pretargeting configuration.

BidderPretargetingConfigAddTargetedAppCall
Adds targeted apps to the pretargeting configuration.

BidderPretargetingConfigAddTargetedPublisherCall
Adds targeted publishers to the pretargeting config.

BidderPretargetingConfigAddTargetedSiteCall
Adds targeted sites to the pretargeting configuration.

BidderPretargetingConfigCreateCall
Creates a pretargeting configuration. A pretargeting configuration's state (PretargetingConfig.state) is active upon creation, and it will start to affect traffic shortly after. A bidder may create a maximum of 10 pretargeting configurations. Attempts to exceed this maximum results in a 400 bad request error.

BidderPretargetingConfigDeleteCall
Deletes a pretargeting configuration.

BidderPretargetingConfigGetCall
Gets a pretargeting configuration.

BidderPretargetingConfigListCall
Lists all pretargeting configurations for a single bidder.

BidderPretargetingConfigPatchCall
Updates a pretargeting configuration.

BidderPretargetingConfigRemoveTargetedAppCall
Removes targeted apps from the pretargeting configuration.

BidderPretargetingConfigRemoveTargetedPublisherCall
Removes targeted publishers from the pretargeting config.

BidderPretargetingConfigRemoveTargetedSiteCall
Removes targeted sites from the pretargeting configuration.

BidderPretargetingConfigSuspendCall
Suspends a pretargeting configuration.

Buyer
RTB Buyer account information.

BuyerCreativeCreateCall
Creates a creative.

BuyerCreativeGetCall
Gets a creative.

BuyerCreativeListCall
Lists creatives.

BuyerCreativePatchCall
Updates a creative.

BuyerGetCall
Gets a buyer account by its name.

BuyerGetRemarketingTagCall
Gets remarketing tag for a buyer. A remarketing tag is a piece of JavaScript code that can be placed on a web page. When a user visits a page containing a remarketing tag, Google adds the user to a user list.

BuyerListCall
Lists all buyer account information the calling buyer user or service account is permissioned to manage.

BuyerMethods
A builder providing access to all methods supported on buyer resources. It is not used directly, but through the RealTimeBidding hub.

BuyerUserListCloseCall
Change the status of a user list to CLOSED. This prevents new users from being added to the user list.

BuyerUserListCreateCall
Create a new user list.

BuyerUserListGetCall
Gets a user list by its name.

BuyerUserListGetRemarketingTagCall
Gets remarketing tag for a buyer. A remarketing tag is a piece of JavaScript code that can be placed on a web page. When a user visits a page containing a remarketing tag, Google adds the user to a user list.

BuyerUserListListCall
Lists the user lists visible to the current user.

BuyerUserListOpenCall
Change the status of a user list to OPEN. This allows new users to be added to the user list.

BuyerUserListUpdateCall
Update the given user list. Only user lists with URLRestrictions can be updated.

CloseUserListRequest
A request to close a specified user list.

Creative
A creative and its classification data.

CreativeDimensions
The dimensions of a creative. This applies to only HTML and Native creatives.

CreativeServingDecision
Top level status and detected attributes of a creative.

Date
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: * A full date, with non-zero year, month, and day values * A month and day value, with a zero year, such as an anniversary * A year on its own, with zero month and day values * A year and month value, with a zero day, such as a credit card expiration date Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

DestinationNotCrawlableEvidence
Evidence that the creative's destination URL was not crawlable by Google.

DestinationNotWorkingEvidence
Evidence of the creative's destination URL not functioning properly or having been incorrectly set up.

DestinationUrlEvidence
The full landing page URL of the destination.

DomainCallEvidence
Number of HTTP calls made by a creative, broken down by domain.

DomainCalls
The number of HTTP calls made to the given domain.

DownloadSizeEvidence
Total download size and URL-level download size breakdown for resources in a creative.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Endpoint
Bidder endpoint that receives bid requests.

GetRemarketingTagResponse
Response for a request to get remarketing tag.

HtmlContent
HTML content for a creative.

HttpCallEvidence
HTTP calls made by a creative that resulted in policy violations.

HttpCookieEvidence
Evidence for HTTP cookie-related policy violations.

Image
An image resource. You may provide a larger image than was requested, so long as the aspect ratio is preserved.

ListBiddersResponse
A response containing bidders.

ListBuyersResponse
A response containing buyer account information.

ListCreativesResponse
A response for listing creatives.

ListEndpointsResponse
A response containing bidder endpoints.

ListPretargetingConfigsResponse
A response containing pretargeting configurations.

ListUserListsResponse
The list user list response.

MediaFile
Information about each media file in the VAST.

NativeContent
Native content for a creative.

NumericTargetingDimension
Generic targeting used for targeting dimensions that contain a list of included and excluded numeric IDs used in app, user list, geo, and vertical id targeting.

OpenUserListRequest
A request to open a specified user list.

PolicyCompliance
Policy compliance of the creative for a transaction type or a region.

PolicyTopicEntry
Each policy topic entry will represent a violation of a policy topic for a creative, with the policy topic information and optional evidence for the policy violation.

PolicyTopicEvidence
Evidence associated with a policy topic entry.

PretargetingConfig
Pretargeting configuration: a set of targeting dimensions applied at the pretargeting stage of the RTB funnel. These control which inventory a bidder will receive bid requests for.

RealTimeBidding
Central instance to access all RealTimeBidding related resource activities

RemoveTargetedAppsRequest
A request to stop targeting the provided apps in a specific pretargeting configuration. The pretargeting configuration itself specifies how these apps are targeted. in PretargetingConfig.appTargeting.mobileAppTargeting.

RemoveTargetedPublishersRequest
A request to stop targeting publishers in a specific configuration. The pretargeting configuration itself specifies how these publishers are targeted in PretargetingConfig.publisherTargeting.

RemoveTargetedSitesRequest
A request to stop targeting sites in a specific pretargeting configuration. The pretargeting configuration itself specifies how these sites are targeted in PretargetingConfig.webTargeting.

StringTargetingDimension
Generic targeting with string values used in app, website and publisher targeting.

SuspendPretargetingConfigRequest
A request to suspend a pretargeting configuration. Sets the configuration's state to SUSPENDED.

UrlDownloadSize
The URL-level breakdown for the download size.

UrlRestriction
Represents the URL restriction (for the URL captured by the pixel callback) for a user list.

UserList
Represents an Authorized Buyers user list. Authorized Buyers can create/update/list user lists. Once a user list is created in the system, Authorized Buyers can add users to the user list using the bulk uploader API. Alternatively, users can be added by hosting a tag on the advertiser's page.

VideoContent
Video content for a creative.

VideoMetadata
Video metadata for a creative.

WatchCreativesRequest
A request to receive push notifications when any of the creatives belonging to the bidder changes status.

WatchCreativesResponse
A response for the request to receive push notification when a bidder's creatives change status.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.