Struct google_realtimebidding1::api::BidderMethods

pub struct BidderMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on bidder resources. It is not used directly, but through the RealTimeBidding hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_realtimebidding1 as realtimebidding1;
 
use std::default::Default;
use oauth2;
use realtimebidding1::RealTimeBidding;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = RealTimeBidding::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `creatives_list(...)`, `creatives_watch(...)`, `endpoints_get(...)`, `endpoints_list(...)`, `get(...)`, `list(...)`, `pretargeting_configs_activate(...)`, `pretargeting_configs_add_targeted_apps(...)`, `pretargeting_configs_add_targeted_publishers(...)`, `pretargeting_configs_add_targeted_sites(...)`, `pretargeting_configs_create(...)`, `pretargeting_configs_delete(...)`, `pretargeting_configs_get(...)`, `pretargeting_configs_list(...)`, `pretargeting_configs_patch(...)`, `pretargeting_configs_remove_targeted_apps(...)`, `pretargeting_configs_remove_targeted_publishers(...)`, `pretargeting_configs_remove_targeted_sites(...)` and `pretargeting_configs_suspend(...)`
// to build up your call.
let rb = hub.bidders();

Implementations

impl<'a, C> BidderMethods<'a, C>

pub fn creatives_list(&self, parent: &str) -> BidderCreativeListCall<'a, C>

Create a builder to help you perform the following task:

Lists creatives.

Arguments

parent - Required. Name of the parent buyer that owns the creatives. The pattern for this resource is either buyers/{buyerAccountId} or bidders/{bidderAccountId}. For buyers/{buyerAccountId}, the buyerAccountId can be one of the following: 1. The ID of the buyer that is accessing their own creatives. 2. The ID of the child seat buyer under a bidder account. So for listing creatives pertaining to the child seat buyer (456) under bidder account (123), you would use the pattern: buyers/456. 3. The ID of the bidder itself. So for listing creatives pertaining to bidder (123), you would use buyers/123. If you want to access all creatives pertaining to both the bidder and all of its child seat accounts, you would use bidders/{bidderAccountId}, e.g., for all creatives pertaining to bidder (123), use bidders/123.

pub fn creatives_watch(
&self,
request: WatchCreativesRequest,
parent: &str
) -> BidderCreativeWatchCall<'a, C>

Create a builder to help you perform the following task:

Watches all creatives pertaining to a bidder. It is sufficient to invoke this endpoint once per bidder. A Pub/Sub topic will be created and notifications will be pushed to the topic when any of the bidder's creatives change status. All of the bidder's service accounts will have access to read from the topic. Subsequent invocations of this method will return the existing Pub/Sub configuration.

Arguments

request - No description provided.

parent - Required. To watch all creatives pertaining to the bidder and all its child seat accounts, the bidder must follow the pattern bidders/{bidderAccountId}.

pub fn endpoints_get(&self, name: &str) -> BidderEndpointGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a bidder endpoint by its name.

Arguments

name - Required. Name of the bidder endpoint to get. Format: bidders/{bidderAccountId}/endpoints/{endpointId}

pub fn endpoints_list(&self, parent: &str) -> BidderEndpointListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the bidder's endpoints.

Arguments

parent - Required. Name of the bidder whose endpoints will be listed. Format: bidders/{bidderAccountId}

pub fn pretargeting_configs_activate(
&self,
request: ActivatePretargetingConfigRequest,
name: &str
) -> BidderPretargetingConfigActivateCall<'a, C>

Create a builder to help you perform the following task:

Activates a pretargeting configuration.

Arguments

request - No description provided.

name - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_add_targeted_apps(
&self,
request: AddTargetedAppsRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigAddTargetedAppCall<'a, C>

Create a builder to help you perform the following task:

Adds targeted apps to the pretargeting configuration.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_add_targeted_publishers(
&self,
request: AddTargetedPublishersRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigAddTargetedPublisherCall<'a, C>

Create a builder to help you perform the following task:

Adds targeted publishers to the pretargeting config.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_add_targeted_sites(
&self,
request: AddTargetedSitesRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigAddTargetedSiteCall<'a, C>

Create a builder to help you perform the following task:

Adds targeted sites to the pretargeting configuration.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_create(
&self,
request: PretargetingConfig,
parent: &str
) -> BidderPretargetingConfigCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a pretargeting configuration. A pretargeting configuration's state (PretargetingConfig.state) is active upon creation, and it will start to affect traffic shortly after. A bidder may create a maximum of 10 pretargeting configurations. Attempts to exceed this maximum results in a 400 bad request error.

Arguments

request - No description provided.

parent - Required. Name of the bidder to create the pretargeting configuration for. Format: bidders/{bidderAccountId}

pub fn pretargeting_configs_delete(
&self,
name: &str
) -> BidderPretargetingConfigDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a pretargeting configuration.

Arguments

name - Required. The name of the pretargeting configuration to delete. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_get(
&self,
name: &str
) -> BidderPretargetingConfigGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a pretargeting configuration.

Arguments

name - Required. Name of the pretargeting configuration to get. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_list(
&self,
parent: &str
) -> BidderPretargetingConfigListCall<'a, C>

Create a builder to help you perform the following task:

Lists all pretargeting configurations for a single bidder.

Arguments

parent - Required. Name of the bidder whose pretargeting configurations will be listed. Format: bidders/{bidderAccountId}

pub fn pretargeting_configs_patch(
&self,
request: PretargetingConfig,
name: &str
) -> BidderPretargetingConfigPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a pretargeting configuration.

Arguments

request - No description provided.

name - Output only. Name of the pretargeting configuration that must follow the pattern bidders/{bidder_account_id}/pretargetingConfigs/{config_id}

pub fn pretargeting_configs_remove_targeted_apps(
&self,
request: RemoveTargetedAppsRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigRemoveTargetedAppCall<'a, C>

Create a builder to help you perform the following task:

Removes targeted apps from the pretargeting configuration.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_remove_targeted_publishers(
&self,
request: RemoveTargetedPublishersRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigRemoveTargetedPublisherCall<'a, C>

Create a builder to help you perform the following task:

Removes targeted publishers from the pretargeting config.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_remove_targeted_sites(
&self,
request: RemoveTargetedSitesRequest,
pretargeting_config: &str
) -> BidderPretargetingConfigRemoveTargetedSiteCall<'a, C>

Create a builder to help you perform the following task:

Removes targeted sites from the pretargeting configuration.

Arguments

request - No description provided.

pretargetingConfig - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn pretargeting_configs_suspend(
&self,
request: SuspendPretargetingConfigRequest,
name: &str
) -> BidderPretargetingConfigSuspendCall<'a, C>

Create a builder to help you perform the following task:

Suspends a pretargeting configuration.

Arguments

request - No description provided.

name - Required. The name of the pretargeting configuration. Format: bidders/{bidderAccountId}/pretargetingConfig/{configId}

pub fn get(&self, name: &str) -> BidderGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a bidder account by its name.

Arguments

name - Required. Name of the bidder to get. Format: bidders/{bidderAccountId}

pub fn list(&self) -> BidderListCall<'a, C>

Create a builder to help you perform the following task:

Lists all the bidder accounts that belong to the caller.