Struct google_realtimebidding1::api::CreativeServingDecision

pub struct CreativeServingDecision {
    pub ad_technology_providers: Option<AdTechnologyProviders>,
    pub china_policy_compliance: Option<PolicyCompliance>,
    pub deals_policy_compliance: Option<PolicyCompliance>,
    pub detected_advertisers: Option<Vec<AdvertiserAndBrand>>,
    pub detected_attributes: Option<Vec<String>>,
    pub detected_click_through_urls: Option<Vec<String>>,
    pub detected_domains: Option<Vec<String>>,
    pub detected_languages: Option<Vec<String>>,
    pub detected_product_categories: Option<Vec<i32>>,
    pub detected_sensitive_categories: Option<Vec<i32>>,
    pub detected_vendor_ids: Option<Vec<i32>>,
    pub last_status_update: Option<String>,
    pub network_policy_compliance: Option<PolicyCompliance>,
    pub platform_policy_compliance: Option<PolicyCompliance>,
    pub russia_policy_compliance: Option<PolicyCompliance>,
}

Top level status and detected attributes of a creative.

This type is not used in any activity, and only used as part of another schema.

Fields
ad_technology_providers: Option<AdTechnologyProviders>

The detected ad technology providers.

china_policy_compliance: Option<PolicyCompliance>

The policy compliance of this creative in China. When approved or disapproved, this applies to both deals and open auction in China. When pending review, this creative is allowed to serve for deals but not for open auction.

deals_policy_compliance: Option<PolicyCompliance>

Policy compliance of this creative when bidding on Programmatic Guaranteed and Preferred Deals (outside of Russia and China).

detected_advertisers: Option<Vec<AdvertiserAndBrand>>

Detected advertisers and brands.

detected_attributes: Option<Vec<String>>

Publisher-excludable attributes that were detected for this creative. Can be used to filter the response of the creatives.list method. If the excluded_attribute field of a bid request contains one of the attributes that were declared or detected for a given creative, and a bid is submitted with that creative, the bid will be filtered before the auction.

detected_click_through_urls: Option<Vec<String>>

The set of detected destination URLs for the creative. Can be used to filter the response of the creatives.list method.

detected_domains: Option<Vec<String>>

The detected domains for this creative.

detected_languages: Option<Vec<String>>

The detected languages for this creative. The order is arbitrary. The codes are 2 or 5 characters and are documented at https://developers.google.com/adwords/api/docs/appendix/languagecodes. Can be used to filter the response of the creatives.list method.

detected_product_categories: Option<Vec<i32>>

Detected product categories, if any. See the ad-product-categories.txt file in the technical documentation for a list of IDs. Can be used to filter the response of the creatives.list method.

detected_sensitive_categories: Option<Vec<i32>>

Detected sensitive categories, if any. Can be used to filter the response of the creatives.list method. See the ad-sensitive-categories.txt file in the technical documentation for a list of IDs. You should use these IDs along with the excluded-sensitive-category field in the bid request to filter your bids.

detected_vendor_ids: Option<Vec<i32>>

IDs of the ad technology vendors that were detected to be used by this creative. See https://storage.googleapis.com/adx-rtb-dictionaries/vendors.txt for possible values. Can be used to filter the response of the creatives.list method. If the allowed_vendor_type field of a bid request does not contain one of the vendor type IDs that were declared or detected for a given creative, and a bid is submitted with that creative, the bid will be filtered before the auction.

last_status_update: Option<String>

The last time the creative status was updated. Can be used to filter the response of the creatives.list method.

network_policy_compliance: Option<PolicyCompliance>

Policy compliance of this creative when bidding in open auction, private auction, or auction packages (outside of Russia and China).

platform_policy_compliance: Option<PolicyCompliance>

Policy compliance of this creative when bidding in Open Bidding (outside of Russia and China). For the list of platform policies, see: https://support.google.com/platformspolicy/answer/3013851.

russia_policy_compliance: Option<PolicyCompliance>

The policy compliance of this creative in Russia. When approved or disapproved, this applies to both deals and open auction in Russia. When pending review, this creative is allowed to serve for deals but not for open auction.