Struct google_realtimebidding1::api::PretargetingConfig

pub struct PretargetingConfig {
    pub allowed_user_targeting_modes: Option<Vec<String>>,
    pub app_targeting: Option<AppTargeting>,
    pub billing_id: Option<String>,
    pub display_name: Option<String>,
    pub excluded_content_label_ids: Option<Vec<String>>,
    pub geo_targeting: Option<NumericTargetingDimension>,
    pub included_creative_dimensions: Option<Vec<CreativeDimensions>>,
    pub included_environments: Option<Vec<String>>,
    pub included_formats: Option<Vec<String>>,
    pub included_languages: Option<Vec<String>>,
    pub included_mobile_operating_system_ids: Option<Vec<String>>,
    pub included_platforms: Option<Vec<String>>,
    pub included_user_id_types: Option<Vec<String>>,
    pub interstitial_targeting: Option<String>,
    pub invalid_geo_ids: Option<Vec<String>>,
    pub maximum_qps: Option<String>,
    pub minimum_viewability_decile: Option<i32>,
    pub name: Option<String>,
    pub publisher_targeting: Option<StringTargetingDimension>,
    pub state: Option<String>,
    pub user_list_targeting: Option<NumericTargetingDimension>,
    pub vertical_targeting: Option<NumericTargetingDimension>,
    pub web_targeting: Option<StringTargetingDimension>,
}

Pretargeting configuration: a set of targeting dimensions applied at the pretargeting stage of the RTB funnel. These control which inventory a bidder will receive bid requests for.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

pretargeting configs activate bidders (response)

pretargeting configs add targeted apps bidders (response)

pretargeting configs add targeted publishers bidders (response)

pretargeting configs add targeted sites bidders (response)

pretargeting configs create bidders (request|response)

pretargeting configs get bidders (response)

pretargeting configs patch bidders (request|response)

pretargeting configs remove targeted apps bidders (response)

pretargeting configs remove targeted publishers bidders (response)

pretargeting configs remove targeted sites bidders (response)

pretargeting configs suspend bidders (response)

Fields
allowed_user_targeting_modes: Option<Vec<String>>

Targeting modes included by this configuration. A bid request must allow all the specified targeting modes. An unset value allows all bid requests to be sent, regardless of which targeting modes they allow.

app_targeting: Option<AppTargeting>

Targeting on a subset of app inventory. If APP is listed in targeted_environments, the specified targeting is applied. A maximum of 30,000 app IDs can be targeted. An unset value for targeting allows all app-based bid requests to be sent. Apps can either be targeting positively (bid requests will be sent only if the destination app is listed in the targeting dimension) or negatively (bid requests will be sent only if the destination app is not listed in the targeting dimension).

billing_id: Option<String>

Output only. The identifier that corresponds to this pretargeting configuration that helps buyers track and attribute their spend across their own arbitrary divisions. If a bid request matches more than one configuration, the buyer chooses which billing_id to attribute each of their bids.

display_name: Option<String>

The diplay name associated with this configuration. This name must be unique among all the pretargeting configurations a bidder has.

excluded_content_label_ids: Option<Vec<String>>

The sensitive content category label IDs excluded in this configuration. Bid requests for inventory with any of the specified content label IDs will not be sent. Refer to this file https://storage.googleapis.com/adx-rtb-dictionaries/content-labels.txt for category IDs.

geo_targeting: Option<NumericTargetingDimension>

The geos included or excluded in this configuration defined in https://storage.googleapis.com/adx-rtb-dictionaries/geo-table.csv

included_creative_dimensions: Option<Vec<CreativeDimensions>>

Creative dimensions included by this configuration. Only bid requests eligible for at least one of the specified creative dimensions will be sent. An unset value allows all bid requests to be sent, regardless of creative dimension.

included_environments: Option<Vec<String>>

Environments that are being included. Bid requests will not be sent for a given environment if it is not included. Further restrictions can be applied to included environments to target only a subset of its inventory. An unset value includes all environments.

included_formats: Option<Vec<String>>

Creative formats included by this configuration. Only bid requests eligible for at least one of the specified creative formats will be sent. An unset value will allow all bid requests to be sent, regardless of format.

included_languages: Option<Vec<String>>

The languages included in this configuration, represented by their language code. See https://developers.google.com/adwords/api/docs/appendix/languagecodes.

included_mobile_operating_system_ids: Option<Vec<String>>

The mobile operating systems included in this configuration as defined in https://storage.googleapis.com/adx-rtb-dictionaries/mobile-os.csv

included_platforms: Option<Vec<String>>

The platforms included by this configration. Bid requests for devices with the specified platform types will be sent. An unset value allows all bid requests to be sent, regardless of platform.

included_user_id_types: Option<Vec<String>>

User identifier types included in this configuration. At least one of the user identifier types specified in this list must be available for the bid request to be sent.

interstitial_targeting: Option<String>

The interstitial targeting specified for this configuration. The unset value will allow bid requests to be sent regardless of whether they are for interstitials or not.

invalid_geo_ids: Option<Vec<String>>

Output only. Existing included or excluded geos that are invalid. Previously targeted geos may become invalid due to privacy restrictions.

maximum_qps: Option<String>

The maximum QPS threshold for this configuration. The bidder should receive no more than this number of bid requests matching this configuration per second across all their bidding endpoints among all trading locations. Further information available at https://developers.google.com/authorized-buyers/rtb/peer-guide

minimum_viewability_decile: Option<i32>

The targeted minimum viewability decile, ranging in values [0, 10]. A value of 5 means that the configuration will only match adslots for which we predict at least 50% viewability. Values > 10 will be rounded down to 10. An unset value or a value of 0 indicates that bid requests will be sent regardless of viewability.

name: Option<String>

Output only. Name of the pretargeting configuration that must follow the pattern bidders/{bidder_account_id}/pretargetingConfigs/{config_id}

publisher_targeting: Option<StringTargetingDimension>

Targeting on a subset of publisher inventory. Publishers can either be targeted positively (bid requests will be sent only if the publisher is listed in the targeting dimension) or negatively (bid requests will be sent only if the publisher is not listed in the targeting dimension). A maximum of 10,000 publisher IDs can be targeted. Publisher IDs are found in ads.txt / app-ads.txt and in bid requests in the BidRequest.publisher_id field on the Google RTB protocol or the BidRequest.site.publisher.id / BidRequest.app.publisher.id field on the OpenRTB protocol.

state: Option<String>

Output only. The state of this pretargeting configuration.

user_list_targeting: Option<NumericTargetingDimension>

The remarketing lists included or excluded in this configuration as defined in UserList.

vertical_targeting: Option<NumericTargetingDimension>

The verticals included or excluded in this configuration as defined in https://developers.google.com/authorized-buyers/rtb/downloads/publisher-verticals

web_targeting: Option<StringTargetingDimension>

Targeting on a subset of site inventory. If WEB is listed in included_environments, the specified targeting is applied. A maximum of 50,000 site URLs can be targeted. An unset value for targeting allows all web-based bid requests to be sent. Sites can either be targeting positively (bid requests will be sent only if the destination site is listed in the targeting dimension) or negatively (bid requests will be sent only if the destination site is not listed in the pretargeting configuration).